refactor(*): changed structure to reflect that of a newly generated angular lib
