--- .travis.yml
@@ -1,24 +1,0 @@
-language: node_js
-cache:
-  yarn: true
-  directories:
-    - ~/.npm
-    - node_modules
-notifications:
-  email: true
-node_js:
-  - "9.3.0"
-# before_install:
-#   - export CHROME_BIN=chromium-browser
-# before_script:
-#   - "export CHROME_BIN=chromium-browser"
-#   - "export DISPLAY=:99.0"
-#   - "sh -e /etc/init.d/xvfb start"
-#   - sleep 3 # give xvfb some time to start
-script:
-  - yarn test
-after_success:
-  - npm run travis-deploy-once "npm run semantic-release"
-branches:
-  only:
-    - master
