+++ package.json
@@ -32,66 +32,9 @@
     "access": "public"
   },
   "peerDependencies": {
-    "moment": "^2.20.1",
-    "jssha": "^2.3.1"
-  },
-  "devDependencies": {
-    "@angular-devkit/schematics": "^0.6.8",
-    "@angular/common": "6.0.4",
-    "@angular/compiler": "6.0.4",
-    "@angular/compiler-cli": "6.0.4",
-    "@angular/core": "6.0.4",
-    "@angular/forms": "6.0.4",
-    "@angular/http": "6.0.4",
-    "@angular/platform-browser": "6.0.4",
-    "@angular/platform-browser-dynamic": "6.0.4",
-    "@types/bootstrap": "^3.3.38",
-    "@types/ckeditor": "0.0.45",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/jquery": "^3.3.0",
-    "@types/jssha": "0.0.29",
-    "@types/moment-range": "^3.0.5",
-    "@types/node": "~6.0.60",
-    "@types/toastr": "^2.1.35",
-    "camelcase": "^4.1.0",
-    "canonical-path": "0.0.2",
-    "codecov": "3.0.0",
-    "commitizen": "2.9.6",
-    "concurrently": "3.5.0",
-    "core-js": "2.5.2",
-    "cz-conventional-changelog": "2.1.0",
-    "glob": "^7.1.2",
-    "gulp": "3.9.1",
-    "gulp-inline-ng2-template": "4.0.0",
-    "husky": "^0.14.3",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
+    "moment": "^2.22.2",
     "jssha": "^2.3.1",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage": "^1.1.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-junit-reporter": "^1.2.0",
-    "karma-mocha-reporter": "^2.2.5",
-    "karma-phantomjs-launcher": "1.0.4",
-    "moment": "^2.20.1",
-    "moment-range": "^3.1.1",
-    "moment-timezone": "^0.5.14",
-    "rollup": "^0.56.1",
-    "rollup-plugin-commonjs": "^8.3.0",
-    "rollup-plugin-node-resolve": "^3.0.3",
-    "rollup-plugin-sourcemaps": "^0.4.2",
-    "rollup-plugin-uglify": "^3.0.0",
-    "rxjs": "6.2.0",
-    "semantic-release": "^12.4.1",
-    "standard-version": "^4.3.0",
-    "systemjs": "0.19.39",
-    "typescript": "~2.7.2",
-    "zone.js": "0.8.26"
+    "@types/jssha": "^0.0.29"
   },
   "dependencies": {}
 }
