+++ src/lib/shared/api-old/api-old-auth-service-config.ts
@@ -17,5 +17,5 @@ export class ApiOldAuthServiceConfig implements AuthServiceConfig {
 
   constructor() {}
 
-  getUri?(path: string){}
+  getUri(path: string){}
 }
