+++ src/lib/shared/api-old/api-old.module.ts
@@ -1,7 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { AuthServiceConfig, CoreAuthServiceConfig } from '../index';
+import { AuthServiceConfig } from '../shared/auth-service-config';
+import { CoreAuthServiceConfig } from '../core-auth-service-config';
 
 import { ApiOldApiService } from './api-old-api.service';
 import { ApiOldAuthService } from './api-old-auth.service';
