--- src/lib/shared/api-old/index.ts
@@ -1,4 +1,0 @@
-export {ApiOldApiService} from './api-old-api.service';
-export {ApiOldAuthService,AuthServiceLogin} from './api-old-auth.service';
-export {ApiOldAuthServiceConfig} from './api-old-auth-service-config';
-export {ApiOldModule} from './api-old.module';
