+++ src/lib/shared/index.ts
@@ -18,5 +18,10 @@ export { AuthService } from './shared/auth.service';
 export { Credentials } from './shared/credentials';
 export { CoreApiSelectorFilter } from './core-api-selector-filter';
 export { CoreApiSelectorParams } from './core-api-selector-params';
-export * from './api-old/index';
-export * from './reporting/index';
+export { ApiOldApiService } from './api-old/api-old-api.service';
+export { ApiOldAuthService, AuthServiceLogin } from './api-old/api-old-auth.service';
+export { ApiOldAuthServiceConfig } from './api-old/api-old-auth-service-config';
+export { ApiOldModule } from './api-old/api-old.module';
+export { ReportingApiService } from './reporting/reporting-api.service';
+export { ReportingAuthService } from './reporting/reporting-auth.service';
+export { ReportingAuthServiceConfig } from './reporting/reporting-auth-service-config';
