--- src/lib/shared/reporting/index.ts
@@ -1,3 +1,0 @@
-export { ReportingApiService } from './reporting-api.service';
-export { ReportingAuthService } from './reporting-auth.service';
-export { ReportingAuthServiceConfig } from './reporting-auth-service-config';
