+++ src/public_api.ts
@@ -1,1 +1,5 @@
+/*
+ * Public API Surface of core
+ */
+
 export * from './lib/index';
