--- tsconfig-spec.json
@@ -1,47 +1,0 @@
-{
-  "compileOnSave": false,
-  "compilerOptions": {
-    "outDir": "./dist",
-    "baseUrl": ".",
-    "rootDir": "./build",
-    "target": "es5",
-    "sourceMap": true,
-    "module": "umd",
-    "moduleResolution": "node",
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "declaration": false,
-    "types": [
-      "jasmine",
-      "toastr",
-      "jssha"
-    ],
-    "typeRoots": [
-      "node_modules/@types"
-    ],
-    "lib": [
-      "es2016",
-      "dom",
-      "es2015"
-    ],
-    "paths": {
-      "/testing": ["./src/testing"]
-    }
-  },
-  "include": [
-    "./build/testing/index.ts",
-    "./build/datetime.ts",
-    "./build/browser-test-shim.ts",
-    "./build/systemjs.config.ts",
-    "./build/**/*.d.ts",
-    "./build/**/*.spec.ts"
-  ],
-  "exclude": [
-    "**/node_modules"
-  ],
-  "angularCompilerOptions": {
-    "annotateForClosureCompiler": true,
-    "strictMetadataEmit": true,
-    "skipTemplateCodegen": true
-  }
-}
