--- tsconfig-spec.upgrade.json
@@ -1,48 +1,0 @@
-{
-    "compileOnSave": false,
-    "compilerOptions": {
-      "outDir": "./dist",
-      "baseUrl": ".",
-      "rootDir": "./src",
-      "target": "es5",
-      "sourceMap": true,
-      "module": "umd",
-      "moduleResolution": "node",
-      "emitDecoratorMetadata": true,
-      "experimentalDecorators": true,
-      "declaration": false,
-      "types": [
-        "jasmine",
-        "toastr",
-        "jssha"
-      ],
-      "typeRoots": [
-        "node_modules/@types"
-      ],
-      "lib": [
-        "es2016",
-        "dom",
-        "es2015"
-      ],
-      "paths": {
-        "/testing": ["./src/testing"]
-      }
-    },
-    "include": [
-      "./src/testing/index.ts",
-      "./src/datetime.ts",
-      "./src/browser-test-shim.ts",
-      "./src/systemjs.config.ts",
-      "./src/**/*.d.ts",
-      "./src/**/*.spec.ts"
-    ],
-    "exclude": [
-      "**/node_modules"
-    ],
-    "angularCompilerOptions": {
-      "annotateForClosureCompiler": true,
-      "strictMetadataEmit": true,
-      "skipTemplateCodegen": true
-    }
-  }
-
